# opt: Replace left joins on foreign keys with inner joins