This change updates the SimplifyLeftJoinWithFilters matcher
to also match cases where we are left joining from an outgoing
foreign key relation i.e. we're always guaranteed to have a match
in the other table.

Fixes #29217

Release note: None