@itsbilal 2018-09-19
Ran the tpcc query with joins (`tpcc-stock-level`), and there was no significant difference in time elapsed in the optbuild phase before and after this change.

Before:
```
BenchmarkPhases/tpcc-stock-level/OptBuild-4      	   30000	     48669 ns/op
BenchmarkPhases/tpcc-stock-level/OptBuild-4      	   30000	     49894 ns/op
BenchmarkPhases/tpcc-stock-level/OptBuild-4      	   30000	     50220 ns/op
```

After:
```
BenchmarkPhases/tpcc-stock-level/OptBuild-4      	   30000	     50262 ns/op
BenchmarkPhases/tpcc-stock-level/OptBuild-4      	   30000	     47391 ns/op
BenchmarkPhases/tpcc-stock-level/OptBuild-4      	   30000	     49674 ns/op
```