@mjibson 2018-09-19
Can you use [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat) to show if there was or wasn't a difference? Usually with at least 5 runs for each.